Renamed checkCorrectPosition as checkPosition
