+++ commandParsers/parserTools/checkPosition.js
